Backend/requirements.txt
flask
werkzeug
PyPDF2
pdf2image
flask-cors
fpdf
google-cloud-vision
google-generativeai
python-dotenv